requirements.txt
motor
aiohttp
requests
python-telegram-bot==20.6
pymongo 
pyrate-limiter==3.1.0
python-telegram-bot[webhooks]
apscheduler==3.6.3
pillow==10.3.0
schedule
pyrogram
tgcrypto 
python-dotenv
cachetools 
telegraph
redis
